requirements.txt
flask
opencv-python-headless
numpy
Pillow
transformers
ultralytics
huggingface_hub
torch
